video: tegra: host: Reorder job completion
Job completion currently first turns off the hardware and then
checks which jobs were completed . This is vulnerable as we
cannot turn off the hardware until jobs have been completed.

This ordering was introduced to avoid race between channel release
and devfreq callbacks. At the time the ordering was selected, it
worked correctly as we were not relying on runtime pm and therefore
we did not start power down sequence at that point.

However, with runtime pm we may start powering down the device at
this point. As we no longer have the race for scaling (it does not
use .deinit()/.init() callbacks) it is safe to reorder this sequence
back to the original form.

Bug 200002762
Bug 1514950

Change-Id: Ida6022b6878d2fa7868ac3369a95240f895fb2ac
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/409595
(cherry picked from commit 51aa2c0294eb9e5fe6b68bfe06700717174b86cd)
Reviewed-on: http://git-master/r/411833
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Sachin Nikam <snikam@nvidia.com>