ARM: tegra: handle touch 3v3 parent rail overcurrent
When touch rail get enabled, there is chances of parent rail to be
short and drop the voltage output of parent rail to be 0V. This can
cause the malfunctioning of system which depends on parent rail.

Add the SW-WAR for enabling touch rail on P1761-A03 designs.
- Enable LDO3 (3.3V) to load switch output.
- Enable load switch and then disable parent rail.

bug 1494740

Change-Id: I283438f421fa72a284bbb8a539c59d0bffee911c
Signed-off-by: Laxman Dewangan <ldewangan@nvidia.com>
Reviewed-on: http://git-master/r/418823
GVS: Gerrit_Virtual_Submit
Reviewed-by: Eric Chuang <echuang@nvidia.com>
Tested-by: Eric Chuang <echuang@nvidia.com>