arm64: lib: optimize unaligned memcpy and copy_*
This version uses only aligned accesses to avoid cache line crossing.
To do this, align the source to be 8 byte aligned, read 8 bytes at a
time, write 1, 2, 4, or 8 bytes to the destination depending on the
destination alignment, and then shift the 8 byte value in the
register to do another store (if needed).

Denver is much faster when reads and writes do not cross a cache line
boundary. Naturally aligned accesses imply that no access crosses a
cache line.

This is at least as fast on A57 as the Linaro version on kernel.org.
But, it is twice as fast on Denver when the source or destination is
not 8 byte aligned.

All loads are hoisted away from their counterpart stores in unrolled
loops to avoid stalls.

copy_*_user is faster on A57 and Denver by a large margin because
this version copies up to 128 bytes per iteration instead of the 8
bytes per iteration maximum on the previous version.

Reuse the same code from all of the functions by using LOAD_USER
and STORE_USER macros. In the memcpy case these do nothing. But,
in the case that they access user memory, they add entries to the
ex_table. Note that the offset is the smallest offset from src or dst
that already has all of the data copied.

bug 1398647

Change-Id: Ie8d3aa746f62f12f33c1340ab670d89378a07270
Signed-off-by: Alex Van Brunt <avanbrunt@nvidia.com>
Reviewed-on: http://git-master/r/590943
Reviewed-on: http://git-master/r/672113
GVS: Gerrit_Virtual_Submit