arm64: tegra21: emc: modify EMC driver init
Change the EMC driver init so that the emc timers - thermal and
periodic training for t21x - happen as a subcall of the general
EMC probe. One caveat is that the thermal device registration
has to be delayed because the EMC driver init runs before the
thermal framework seems to exist.

Change-Id: I37e27374d436bf8fb5c8e502659bdfd249d1878f
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/660451
Reviewed-by: Mitch Luban <mluban@nvidia.com>
GVS: Gerrit_Virtual_Submit