video: tegra: hdmi: Port the hpd_switch to the hdmi2.0 driver
The HDMI driver had the hpd_switch which signals to the userspace
when the HDMI is ready (after EDID parsing). Android currently
uses this switch to figure out when a hot plug has occurred. This
is technically an abuse of the switch api and we should refactor
this to a cleaner implementation later.

Bug 1570417

Change-Id: I4b233fc29a506f576555b1b7e1bacc6b06c818e7
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/594608
(cherry picked from commit 3567d1b4b851bd181a4511e68978e348cbc17c2b)
Reviewed-on: http://git-master/r/602496
Reviewed-by: Jon Mayo <jmayo@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vinayak Pane <vpane@nvidia.com>