net:wireless:bcmdhd: Fix dhd filter for 64bit
dhd filter is not gettig enabled due to size mismatch in driver
and firmware. Driver is returning the 64 bit union size where
as fimrware expects the 32 bit pointer size. Removing the
decrypt_ctx as it is not used and causing the union size
mismatch with firmware.

Bug 200054625

Change-Id: I5017d5e74d5aa9bd56c602ed13f9fc49ea788538
Signed-off-by: Raghavender <kraghavender@nvidia.com>
Reviewed-on: http://git-master/r/606039
Reviewed-by: Narayan Reddy <narayanr@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>