video: tegra: host: Tune actmon for T210
This patch tunes T210 actmon to produce sensible load readings given
the usual tasks; Usually the tasks are relatively small (length of
tens of microseconds) and hence we only increase this slightly to
get better estimate in most use cases.

In addition, make K factor higher to give more weight for the old
measurement results to avoid overshoorting of scaling.

Change-Id: Ibaf66ee93a7971cf50211461d51009f624243069
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/656132
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Mitch Luban <mluban@nvidia.com>