arm64: tegra21: emc: Update to 21018 sequence
Update the 21015 code path to 21018. This incorporates a few changes
to fix sporadic and rare hang issues. The most notable is due to the
DLL sometimes recalibrating at exactly the same time the EMC is
configured to write directly to the active copy of all registers.

Change-Id: Ib24845a15f1d6e47a6ff0ee92b511f7629cb4eb5
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/660452
Reviewed-by: Mitch Luban <mluban@nvidia.com>
GVS: Gerrit_Virtual_Submit