video: tegra: dc: Fix CMU at boot
CBoot currently programs the CMU before handover to the kernel. During
the handover, we are currently seeing a glitch breaking seamless
display. The root cause is because we were turning the CMU off,
programming the CMU and then turning it back on.

This change:

1. Changes the tegra_dc_init() function to call the update_cmu_aligned
instead of the update_cmu. The update_cmu function turns off the CMU,
reprograms it and then turns it back on. Using the aligned call prevents
us from turning the CMU on and off.

2. Splits up the aligned call into a locked and unlocked function. The
dc_init needs the unlocked variation to be called.

3. Adds setting LUT1 to the frame_end callback for the CMU update. The
existing code didn't need to touch LUT1 so the support wasn't there.
However, now that we are using the aligned call to perform a full
update, this does need to be added.

4. Adds a "dc->cmu_shadow_force_update". Setting this to true will cause
the frame_end callback to force update all the CMU registers. Because
the HW initializes these registers as garbage, we need to force an
update, even for the zero'd out values

5. Cleans up the unnecessary overwrite of the cmu_dirty. Given the
changes around the cmu_update function, this shouldn't have been working
from as far as I can tell.

Bug 1587291

Change-Id: I8b563c7a117a9258de9b84c3b5d6eb787c4efe0f
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/660569
GVS: Gerrit_Virtual_Submit
Reviewed-by: Tao Xie <txie@nvidia.com>
Reviewed-by: Jon Mayo <jmayo@nvidia.com>
Reviewed-by: Yu-Huan Hsu <yhsu@nvidia.com>