video: tegra: Reorder ISP and DC initializations
DC is initialized by bootloader with a certain configuration.
Based on debugging this configuration depends on the VE power
partition. Currently we probe ISP before DC and as a result we
observe corruption on the screen due to the configuration left
by bootloader.

In order to avoid corruption during boot, this patch reorders
module initialization so that ISP is initialized after DC.

Bug 1577602

Change-Id: Ibbee920e643389c5328b6466afb379e6d0b9ec1f
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/660822
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Tested-by: Ankita Garg <ankitag@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>
Reviewed-by: Min-wuk Lee <mlee@nvidia.com>