drivers: cpuquiet: use work queue to apply constraints
There may be a deadlock if apply_constraints() is called directly in
min/max_cpus_notify() since apply_constraints() will call cpu_down()
and then need to acquire console lock. The deadlock happens in case
a thread who holds the console lock is blocked on pm_qos_lock which
is in turn held by another thread calling min/max_cpus_notify().

To avoid the deadlock, use a work queue to apply constraints instead
of calling it directly in the notifiers which are called outside of
cpuquiet driver and the dependancy is unpredictable.

Bug 1588484

Change-Id: I64ced9e1749e44958ece0e5ecb5e2c77b952805b
Signed-off-by: Allen Yu <alleny@nvidia.com>
Reviewed-on: http://git-master/r/662729
Reviewed-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Tested-by: David Dastous St Hilaire <ddastoussthi@nvidia.com>
Reviewed-by: David Dastous St Hilaire <ddastoussthi@nvidia.com>
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>