video: tegra: host: Implement shutdown
In rare occassions something starts waiting for a syncpoint
when we are starting shutdown and as a result the shutdown process
may get disrupted when the syncpoint value is not achieved.

This patch adds shutdown routine to host1x. During shutdown we..:
- Release all syncpoint waiters with -ETIMEDOUT error code
- Prevent new requests by disabling pm runtime from host1x

Bug 200064323

Change-Id: If20fcf78af14a3ee639496da2c8f9f8ef43733db
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
(cherry-picked from commit 2d906a8249b3cbe73539160be43156b8305f03e5)
Reviewed-on: http://git-master/r/663958
Reviewed-by: Automatic_Commit_Validation_User