arm64: tegra21: emc: Fix derating support
Copy necessary training parameters from the regular tables to the
derated tables. Also do this for the periodic training parameters
stored in the table after each clock change and after each periodic
training cycle.

Keep track of the PLL currently in use so that when we do derating
we can use that came clock source. This is important for the PLLM
vs PLLMB switching. The tables only specify PLLM. If the current
frequency is running on PLLMB and then you either derate or un-derate
then you will swap to PLLM which is disabled.

Lastly, change the refresh settings to only apply 4x when derating
is not also being applied. The derated tables have changed in that
they now have the 4x refresh component built in for T210.

Change-Id: I849829bdc6ac033e3aeebe3e7877d030b83acc48
Signed-off-by: Alex Waterman <alexw@nvidia.com>
Reviewed-on: http://git-master/r/667368
Reviewed-by: Emad Mir <emir@nvidia.com>
Tested-by: Emad Mir <emir@nvidia.com>