video: tegra: nvmap: corrected procrank PSS calculation
Modified PSS calculation in NvMap "procrank" debugfs node to be
performed similar to PSS calculation in "smaps" procfs node. This is
done by walking page table for VMAs associated with client's handles.

Earlier NvMap procrank used to assume all non-mapped pages as non-PSS
and remaining as PSS, which is not correct when multiple processes
share the same handle.

Bug 1575148

Change-Id: If8477d88b35b86b4054d5b05ff1f05fd66ef7c20
Signed-off-by: Maneet Singh <mmaneetsingh@nvidia.com>
Reviewed-on: http://git-master/r/670524
Reviewed-by: Emad Mir <emir@nvidia.com>
Tested-by: Emad Mir <emir@nvidia.com>