video: tegra: host: Fix actmon thresholds
Actmon thresholds were calculated in assumption that the
threshold load values are scaled from 0 to 255. However,
in T210 (where these interrupts are only used) this is not
the case but we need to take into account the number of
clocks actmon ticks during sample and multiply that by
the clock divider (256 or 6636).

In addition to fixing the divider, this patch explicitly
disabled actmon interrupt after receiving it. The interrupt
is re-enabled after handling the interrupt in worker thread.

Bug 200069284

Change-Id: I9dc13acbe61bf0030f337a108e11e41ddd95f56a
Signed-off-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-on: http://git-master/r/672733
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Seema Khowala <seemaj@nvidia.com>
Tested-by: Seema Khowala <seemaj@nvidia.com>
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
Reviewed-by: Raymond Poudrier <rapoudrier@nvidia.com>