video: tegra: dc: bypass EDID checksum
We have seen some old TVs which actually have a bad checksum hardcoded
in their EDIDs. Given our current use of the checksum code, this causes
things to break. These bad TVs actually have the right modes listed in
their timing blocks. Just the checksum is bad.

The bootloader doesn't fully validate the EDID checksum. This means that
it is actually possible to boot with HDMI connected on these "bad" TVs
(since the kernel just copies over the mode set by BL). However, as soon
as we hot plug in the kernel, we break HDMI. This is because in the current
state of code, if the EDID checksum fails, we assume that to be an equal
to an i2c read failure.

Instead of assuming a failed checksum to be equal to a read failure,
this change adds a retry mechanism built into the EDID read.
Specifically, if we read the EDID multiple number of times and calculate
the checksum to be exactly the same, assume that the TV has a bad
checksum and just ignore it. This change also adds the checksum
calculated from the EDID block in a print since it really helps in
debugging.

This approach is also recommended in CEA-861-F Section F.3.5.

Bug 1608233

Change-Id: I0384c90345a599c328337f7d63362df900a4bf4b
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/718345
Reviewed-by: Jon Mayo <jmayo@nvidia.com>
Reviewed-by: Animesh Kishore <ankishore@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>