drivers: video: tegra: No HDCP in recovery
Some TVs are known to send us HPD when we start HDCP negogitation. This
is a problem for us in the recovery kernel since the expected sequence
of events is that BL should setup the mode, and the kernel should just
copy it for recovery.

However, in the recovery if HDCP is started, TVs HPD us. This becomes a
problem since the recovery kernel has no HPD support and results in a
blank screen.

The previous hack was to disable HPD events in the recovery kernel.
However, some TVs don't like it when we don't disable our output. This
hack adds on top of that and disables HDCP in the recovery kernel,
hoping that the TVs don't HPD us.

Bug  200070455

Change-Id: I00669dbe4e1f4525996f24511825c6a5e0a43162
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/727313
Reviewed-by: Vinayak Pane <vpane@nvidia.com>
Tested-by: Vinayak Pane <vpane@nvidia.com>