drivers: cpuquiet: Plug/unplug CPUs in order
When satisfying the min/max constraints on online cpus, plug/unplug the
CPUs in order to make system behavior more predictable. There is
slightly higher penalty for the max case now as we need to find the last
online CPU but this shouldn't be significant. Also cleanup some code
while at it.

Change-Id: Ief4e4a76bbb960e476d6534bab30118b7326c91f
Signed-off-by: Sai Gurrappadi <sgurrappadi@nvidia.com>
Reviewed-on: http://git-master/r/740425
Tested-by: David Lock <dlock@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ilan Aelion <iaelion@nvidia.com>
Reviewed-by: Vinayak Pane <vpane@nvidia.com>