video: tegra: dc: black background for yuv modes
If a non-fullscreen window is displayed when YUV 420 10bpc
mode is active, frame with special packing pattern must be
displayed on the background to simulate background color.

It is enough to replicate only a single line over the whole
display height. This decreases amount of memory needed to
store the special bit pattern. The memory for the pattern
is allocated separatelly from framebuffer memory.

Bug 1640154

Change-Id: I6e4e8773c896a950edee33a2e3d47f5c75d93279
Signed-off-by: Michael Frydrych <mfrydrych@nvidia.com>
Reviewed-on: http://git-master/r/736785
Reviewed-by: Animesh Kishore <ankishore@nvidia.com>
Reviewed-by: Jon Mayo <jmayo@nvidia.com>
Reviewed-by: Aly Hirani <ahirani@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Mitch Luban <mluban@nvidia.com>