security: tlk_driver: bump ss buf size
To accomodate a larger secure storage read/write max
data size (from 2K to 8K) we need to increase the
size of shared buffer into/from which this data is
transferred.

So switch from 4K to 16K (giving a little more headroom
in case we need more space in the future).

Bug 1649599

Change-Id: I17a0f38097be4b2fc55a1f53a5fa190912cae251
Signed-off-by: Scott Long <scottl@nvidia.com>
Reviewed-on: http://git-master/r/751049
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Tested-by: Sang-Hun Lee <sanlee@nvidia.com>
Reviewed-by: Vinayak Pane <vpane@nvidia.com>