drivers: video: tegra: dc: Fix dpaux seq
This change fixes the dpaux programming sequence. Specifically, the tegrashell
scripts recommend to program the mode before powering on the pad. In the
current state of code, we were turning on the pad first (in the default dpaux
mode) and then switching the mode.

This specifically led to an issue with certain ddc transactions where the
voltage on SDA lines would hover at ~1.0 V (instead of 0 V/5 V high).
Changing this programming sequence helps remove this hover voltage with
certain sinks.

Bug 1652890

Change-Id: I70ef3f351b9fe2a4126f05a0b546e4360365071a
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/754684
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Shu Zhong <shuz@nvidia.com>
Reviewed-by: Mitch Luban <mluban@nvidia.com>
Reviewed-by: Jon Mayo <jmayo@nvidia.com>
Reviewed-by: Animesh Kishore <ankishore@nvidia.com>
Reviewed-by: Vinayak Pane <vpane@nvidia.com>