drivers: video: tegra: Remove EDID clockrate check
This change removes the mode-filtering which used to remove the
modes from the modedb based on the max TMDS clock rate suggested
by the EDID.

From the HDMI 2.0 spec, Section 10.3.2, with regards to
Max_TMDS_Character_Rate: "This field may be set by the Sink to a
value below the TMDS Character Rate corresponding to the maximum
Pixel clock rate at the maximum color depth". HDMI 1.4 has a same
clause for Max_TMDS_Clock.

This means that the actual maximum TMDS rate supported by the TV
can be higher than the value suggested in the EDID. This filter
should have never been added in the first place since it can
remove actually supported modes.

Bug 1653863

Change-Id: I087b7a4efc3e43de4dfb34071adb88dc9cc6b3e1
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/757615
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Manish Tuteja <mtuteja@nvidia.com>