video: tegra: dc: Remove DPAUX powerdown
This removes the explicit dpaux pad powerdown that was being called from
tegra_dc_controller_enable. Dpaux pads are meant to be refcounted using
the ddc_enable and ddc_disable functons. When the refcounts drop to
zero, they are responsible for calling the powerdown function.

For some unknown reason, dc_controller_enable was explicitly powering
down the pads (without going through the refcounting). This meant that
it is possible for the refcounting system to think that the pads are
powered, when in fact they were not.

In rel-22, it seems like it hardcodes the DPAUX (not DPAUX1) pads to be
powered off. This is a bug and we should never be doing this.

Bug 200111886
Bug 1659406
Bug 200114600

Change-Id: I6f3f8028103efbeb9d79af1adc3c5de06da2fd08
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/769599
(cherry picked from commit e182a5e823074eafe1a746d589c4c7dfccf5db54)
Reviewed-on: http://git-master/r/777430
Reviewed-by: Automatic_Commit_Validation_User
GVS: Gerrit_Virtual_Submit
Reviewed-by: Vinayak Pane <vpane@nvidia.com>