video: tegra: hdmi: Fix locks and race condition
The HDMI driver had a couple of race conditions that this change
attempts to fix.

1. ddc_enable/ddc_disable

tegra_hdmi_ddc_enable() and _ddc_disable() functions need to be
called before performing any ddc/i2c transactions on T210. These are
called at least from (1) edid thread, (2) scdc thread, (3) hdcp thread,
(4) from userspace by crc checker. The enable and disable functions
makes sure that the pads are powered on or powered off (with refcounting).
The hdmi driver was using a standard int with no locking around it,
causing typical race conditions and refcounting to go bad. This change
adds a new mutex to protect ddc_refcount.

2. clock_refcount_lock

The hdmi clocks are also refcounted with tegra_hdmi_get() and
tegra_hdmi_put(). They also make sure that the hdmi clocks are only
enabled and configured on the first enable() and turned off on the last
disable(). It was also using a standard int with no refcounting. This
change also adds a new mutex to protect clock_refcount_lock.

3. ddc_lock

ddc_lock was being held across some scdc transactions. However, this
is not needed. The i2c_transfer() function should be responsible for
maintaining the bus lock. Also, this is made worse by the fact that not
all ddc transactions were holding this lock.

Bug 200103101

Change-Id: Idaece90a5d7f9575c3f97e2848bcf1c605de693c
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/766759
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Vinayak Pane <vpane@nvidia.com>