video: tegra: dc: Fix HDMI cert compliance
This change fixes the HDMI cert compliance for the 23.976 Hz mode. In
that mode, we were not setting the extended_vic correctly (it was being
set as 0). This was because the 23.976's fb_videomode didn't match
exactly to the 24 Hz one specified in the hardcoded list.

This is only a short-term WAR to remove the VMODE and fix the pclk
before comparing the fb_videomodes. It also removes the comparison on
the fb_videomode's refresh field (which unfortunately is not accurate at
all due to all the rounding that we've had to do because of 23.976 Hz
modes).

As a temporary WAR, this change replaces the fb_mode_is_equal() call
with an expanded if check. This probably needs be redone before
integration to other branches, but should fix the associated cert
failure.

Bug 200137448

Change-Id: Ica8bfd6c2292acd7911f33fe317dad107ae571c7
Signed-off-by: Aly Hirani <ahirani@nvidia.com>
Reviewed-on: http://git-master/r/796834
GVS: Gerrit_Virtual_Submit
Reviewed-by: Bharat Nihalani <bnihalani@nvidia.com>